+++ poem/src/error.rs
@@ -35,9 +35,6 @@ pub trait ResponseError {
     where
         Self: StdError + Send + Sync + 'static,
     {
-        if self.status() == StatusCode::NOT_MODIFIED {
-            return Response::builder().status(self.status()).finish();
-        }
         Response::builder()
             .status(self.status())
             .body(self.to_string())
@@ -311,6 +308,13 @@ impl Error {
         impl ResponseError for StatusError {
             fn status(&self) -> StatusCode {
                 self.0
+            }
+
+            fn as_response(&self) -> Response
+            where
+                Self: StdError + Send + Sync + 'static,
+            {
+                self.0.into_response()
             }
         }